+++ runtime/logger.js
@@ -9,6 +9,7 @@ exports.ChatLog = new Winston.Logger({
     new (require('winston-daily-rotate-file'))({
       level: 'info',
       filename: __dirname + '/../logs/chatlog',
+      json: false,
       datePattern: '-dd-MM-yyyy.log'
     })
   ]
@@ -19,6 +20,7 @@ exports.DebugModeLog = new Winston.Logger({
     new (require('winston-daily-rotate-file'))({
       level: 'debug',
       filename: __dirname + '/../logs/debugmodelog',
+      json: false,
       datePattern: '-dd-MM-yyyy.log'
     })
   ]
@@ -29,6 +31,7 @@ exports.VerboseModeLog = new Winston.Logger({
     new (require('winston-daily-rotate-file'))({
       level: 'info',
       datePattern: '-dd-MM-yyyy.log',
+      json: false,
       filename: __dirname + '/../logs/verbosemodelog'
     })
   ]